Fix setpoint parsing error (#11635)

+++ bundles/org.openhab.binding.radiothermostat/src/main/java/org/openhab/binding/radiothermostat/internal/dto/RadioThermostatTstatDTO.java
@@ -35,10 +35,10 @@ public class RadioThermostatTstatDTO {
     private Integer programMode;
 
     @SerializedName("t_heat")
-    private Integer heatTarget;
+    private Double heatTarget;
 
     @SerializedName("t_cool")
-    private Integer coolTarget;
+    private Double coolTarget;
 
     @SerializedName("override")
     private Integer override;
@@ -86,19 +86,19 @@ public class RadioThermostatTstatDTO {
         this.programMode = programMode;
     }
 
-    public Integer getHeatTarget() {
+    public Double getHeatTarget() {
         return heatTarget;
     }
 
-    public void setHeatTarget(Integer heatTarget) {
+    public void setHeatTarget(Double heatTarget) {
         this.heatTarget = heatTarget;
     }
 
-    public Integer getCoolTarget() {
+    public Double getCoolTarget() {
         return coolTarget;
     }
 
-    public void setCoolTarget(Integer coolTarget) {
+    public void setCoolTarget(Double coolTarget) {
         this.coolTarget = coolTarget;
     }
 
@@ -129,9 +129,9 @@ public class RadioThermostatTstatDTO {
      */
     public Integer getSetpoint() {
         if (mode == 1) {
-            return heatTarget;
+            return heatTarget.intValue();
         } else if (mode == 2) {
-            return coolTarget;
+            return coolTarget.intValue();
         } else {
             return 0;
         }
